backend/requirements.txt
Flask==3.0.0
flask-cors==4.0.0
google-generativeai==0.3.2
chromadb==0.4.22
PyPDF2==3.0.1
werkzeug==3.0.1
python-dotenv==1.0.0
pytesseract==0.3.10
PyMuPDF==1.23.8
Pillow==10.1.0
